Backend/requirements.txt
flask
flask-cors
transformers
torch
pillow
google-generativeai
firebase-admin
python-dotenv
gunicorn